Add crawler
With this commit, one can run a fairly expensive operation in which the
program starts from Category:India page and gets all the lists directly
or indirectly linked to it and downloads these lists in separate sections.
Some formatting is required to make sure that links point correctly
to Wikipedia.